feat(drawer): new drawer component interface
BREAKING CHANGE: see docs
