+++ docs/src/content/components/position.mdx
@@ -2,8 +2,6 @@
 name: Position
 route: /components/position
 ---
-
-import Position from '@pluralsight/ps-design-system-position'
 import IframeResizer from 'iframe-resizer-react'
 
 # Position
