+++ packages/drawer/src/css/index.ts
@@ -8,105 +8,64 @@ import {
   motion,
   type
 } from '@pluralsight/ps-design-system-core'
-import {
-  defaultName as themeDefaultName,
-  names as themeNames
-} from '@pluralsight/ps-design-system-theme'
-
-import * as vars from '../vars'
+import { names as themeNames } from '@pluralsight/ps-design-system-theme'
 
 export default {
-  [`.psds-drawer.psds-theme--${themeNames.light}`]: {
-    borderTop: `1px solid ${colorsBorder.lowOnLight}`
-  },
-  [`.psds-drawer.psds-theme--${themeDefaultName}`]: {
-    borderTop: `1px solid ${colorsBorder.lowOnDark}`
-  },
-
-  // __base
-  [`.psds-drawer__base`]: {
-    position: 'relative',
-    paddingRight: vars.toggleAreaWidth,
-    cursor: 'pointer',
-    transition: `background ${motion.speedNormal}`,
-
-    [`&:hover`]: {
-      background: colorsBackgroundUtility[25]
-    }
-  },
-  [`.psds-drawer__base--isOpen`]: {
-    background: colorsBackgroundUtility[25],
-
-    [`&:hover`]: {
-      background: colorsBackgroundUtility[25]
-    }
-  },
-
-  // __panel / __panel-content
-  [`.psds-drawer__panel`]: {
-    background: colorsBackgroundDark[2],
-    transition: `box-shadow ${motion.speedNormal}`
-  },
-  [`.psds-drawer__panel.psds-theme--${themeNames.light}`]: {
-    background: colorsBackgroundLight[2],
-    boxShadow: 'inset 0 1px 3px 0 rgba(0,0,0,0.1)'
-  },
-  [`.psds-drawer__panel.psds-theme--${themeDefaultName}`]: {
-    boxShadow: 'inset 0 1px 3px 0 rgba(0,0,0,0.5)'
-  },
-  [`.psds-drawer__panel-content`]: {
-    marginRight: `calc(-1 * ${vars.toggleAreaWidth})`,
-    paddingLeft: layout.spacingMedium,
-    paddingRight: vars.toggleAreaWidth
-  },
-
-  // __toggle-button-container / __toggle-button
-  [`.psds-drawer__toggle-button-container`]: {
-    position: 'absolute',
-    top: 0,
-    right: layout.spacingMedium,
-    bottom: 0,
-    display: 'flex',
-    alignItems: 'center'
-  },
-  [`.psds-drawer__toggle-button`]: {
-    height: '24px',
-    overflow: 'hidden',
-    fontSize: type.fontSizeXSmall,
-    padding: 0,
-    cursor: 'pointer',
-    border: 'none',
-    background: 'none',
-    transition: `all ${motion.speedNormal}`
-  },
-  [`.psds-drawer__toggle-button.psds-theme--${themeNames.light}`]: {
+  [`.psds-drawer__summary.psds-theme--${themeNames.light}`]: {
+    borderTop: `1px solid ${colorsBorder.lowOnLight}`,
     color: colorsTextIcon.lowOnLight,
-
     [`&:hover, &:active`]: {
       color: colorsTextIcon.highOnLight
     }
   },
-  [`.psds-drawer__toggle-button.psds-theme--${themeDefaultName}`]: {
+  [`.psds-drawer__summary.psds-theme--${themeNames.dark}`]: {
+    borderTop: `1px solid ${colorsBorder.lowOnDark}`,
     color: colorsTextIcon.lowOnDark,
-
     [`&:hover, &:active`]: {
       color: colorsTextIcon.highOnDark
     }
   },
-
-  // __rotatable
-  [`.psds-drawer__rotatable`]: {
-    transition: `transform ${motion.speedNormal}`,
-    lineHeight: 0
-  },
-  [`.psds-drawer__rotatable--isOpen`]: {
-    transform: 'rotateZ(180deg)'
-  },
-
-  // __collapsible
-  [`.psds-drawer__collapsible`]: {
+  '.psds-drawer__summary': {
+    display: 'inline-flex',
+    width: '100%',
     overflow: 'hidden',
-    visibility: 'hidden',
-    transition: `height ${motion.speedNormal}`
+    fontSize: type.fontSizeXSmall,
+    cursor: 'pointer',
+    border: 'none',
+    background: 'none',
+    alignItems: 'center',
+    transition: `all ${motion.speedNormal}`,
+    padding: `${layout.spacingSmall} ${layout.spacingMedium}`,
+    [`&:hover`]: {
+      background: colorsBackgroundUtility[25]
+    }
+  },
+  '.psds-drawer__summary.psds-drawer--isOpen': {
+    background: colorsBackgroundUtility[25],
+    [`&:hover`]: {
+      background: colorsBackgroundUtility[25]
+    }
+  },
+  '.psds-drawer__details': {
+    background: colorsBackgroundDark[2],
+    transition: `all ${motion.speedNormal}`
+  },
+  [`.psds-drawer__details.psds-theme--${themeNames.light}`]: {
+    background: colorsBackgroundLight[2],
+    boxShadow: 'inset 0 1px 3px 0 rgba(0,0,0,0.1)'
+  },
+  [`.psds-drawer__details.psds-theme--${themeNames.dark}`]: {
+    boxShadow: 'inset 0 1px 3px 0 rgba(0,0,0,0.5)'
+  },
+  '.psds-drawer__icon-slot': {
+    display: 'inline-block',
+    paddingLeft: layout.spacingSmall,
+    marginLeft: 'auto'
+  },
+  [`.psds-drawer__rotatable`]: {
+    transition: `transform ${motion.speedNormal}`
+  },
+  [`.psds-drawer__rotatable.psds-drawer--isOpen`]: {
+    transform: 'rotateZ(180deg)'
   }
 }
