+++ packages/drawer/src/index.ts
@@ -1,3 +1,3 @@
-export { default } from './react'
+export { default, useDrawerContext } from './react'
 export * from './vars'
 export { default as css } from './css'
